\name{CVThresh}
\alias{CVThresh-package}
\alias{CVThresh}
\docType{package}
\title{Level-Dependent Cross-Validation Approach for Wavelet Thresholding}
\description{
   This package carries out level-dependent
cross-validation method for the selection of thresholding
value in wavelet shrinkage. This procedure is implemented
by coupling a conventional cross validation with an
imputation method due to a limitation of data length,
a power of 2. It can be easily applied to classical
leave-one-out and k-fold cross validation.
Since the procedure is computationally fast,
a level-dependent cross validation can be performed for
wavelet shrinkage of various data such as a data
with correlated errors.
}
\keyword{nonparametric}
