\name{ipd}
\docType{data}
\alias{ipd}
\title{Inductance plethysmography data}
\description{
4,096 observations of
inductance plethysmography data regularly sampled at 50Hz starting at 1229.98 seconds.
}
\usage{
data(ipd)
}
\format{
time series.
}
\source{
This data set contains 4,096 observations of
inductance plethysmography data regularly sampled at 50Hz starting at 1229.98 seconds. The data
were collected in an investigation of the recovery of patients after general anesthesia.

The data set was used in Nason (1996) to illustrate cross-validation method for threshold selection. 
See the reference; Nason, G.P. (1996) Wavelet shrinkage by cross-validation. \emph{Journal of the Royal Statistical
Society Ser. B} \bold{58}, 463--479.
}
\keyword{datasets}
